Move card, input, navigation component classe into their own files
needed as these files will indiviudally grow, better to split the logic now
