+++ src/python/impactx/dashboard/Input/__init__.py
@@ -1,9 +1,14 @@
-from .components.components import CardBase, CardComponents, InputComponents, NavigationComponents
+from .components.card import CardBase, CardComponents
+from .components.input import InputComponents
+from .components.navigation import NavigationComponents
 from .defaults import DashboardDefaults
 from .generalFunctions import generalFunctions
 
 __all__ = [
+    "CardBase",
     "CardComponents",
+    "InputComponents",
+    "NavigationComponents",
     "DashboardDefaults",
     "generalFunctions",
     "InputComponents",
